refactor: server modules with new version route. add: version on ui
